feat: Release feature branch (#360)
BREAKING CHANGE: Storing profiles in config file instead of keytar.

+++ src/commands/profiles/remove.js
@@ -10,7 +10,10 @@ class ProfilesRemove extends TwilioClientCommand {
     const verdict = await this.confirmRemoveProfile();
     if (verdict === true) {
       const keyVerdict = await this.confirmRemoveKey();
-      const credentials = await this.deleteLocalKey(deleteProfile);
+      let credentials;
+      if (!this.userConfig.profiles[deleteProfile.id]) {
+        credentials = await this.deleteLocalKey(deleteProfile);
+      }
       await this.deleteRemoteKey(deleteProfile, keyVerdict, credentials);
       this.logger.info(`Deleted ${deleteProfile.id} profile.`);
       this.userConfig.removeProfile(deleteProfile);
@@ -33,7 +36,7 @@ class ProfilesRemove extends TwilioClientCommand {
         'Are you sure you want to remove the active profile? Run "twilio profiles:use" to set another profile as active.',
       );
     }
-    if (this.userConfig.profiles.length === 1) {
+    if (this.userConfig.projects.length + Object.keys(this.userConfig.profiles).length === 1) {
       this.logger.warn(
         'Are you sure you want to remove the last profile? Run "twilio profiles:create" to create another profile.',
       );
@@ -54,7 +57,8 @@ class ProfilesRemove extends TwilioClientCommand {
   async deleteRemoteKey(profileDelete, keyVerdict, credentials) {
     if (keyVerdict === true) {
       try {
-        await this.twilioClient.api.keys(credentials.apiKey).remove();
+        const apiKey = profileDelete.apiKey || credentials.apiKey;
+        await this.twilioClient.api.keys(apiKey).remove();
         this.logger.info('The API Key has been deleted from The Twilio console.');
       } catch (error) {
         this.logger.error(
